Pagination feature tests

+++ heseya/pagination/config/pagination.php
@@ -2,8 +2,8 @@
 
 return [
 
-    'per_page' => env('PAGINATION_DEFAULT', 24),
+    'per_page' => (int) env('PAGINATION_DEFAULT', 24),
 
-    'max' => env('PAGINATION_MAX', 500),
+    'max' => (int) env('PAGINATION_MAX', 500),
 
 ];
